+++ projects/storefrontlib/src/cms-components/navigation/category-navigation/category-navigation.module.ts
@@ -1,12 +1,11 @@
-import { BootstrapModule } from '../../bootstrap.module';
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-
-import { ConfigModule, CmsConfig, CmsService } from '@spartacus/core';
+import { NgModule } from '@angular/core';
+import { CmsConfig, CmsService, ConfigModule } from '@spartacus/core';
+import { CmsComponentData } from '../../../cms-structure/page/model/cms-component-data';
+import { BootstrapModule } from '../../../lib/bootstrap.module';
+import { NavigationComponentService } from '../navigation/navigation.component.service';
 import { NavigationModule } from '../navigation/navigation.module';
 import { CategoryNavigationComponent } from './category-navigation.component';
-import { NavigationComponentService } from '../navigation/navigation.component.service';
-import { CmsComponentData } from '../../../cms-structure/page/model/cms-component-data';
 
 @NgModule({
   imports: [
