+++ projects/storefrontlib/src/cms-components/navigation/footer-navigation/footer-navigation.module.ts
@@ -3,8 +3,8 @@ import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { CmsConfig, CmsService, ConfigModule } from '@spartacus/core';
 import { CmsComponentData } from '../../../cms-structure/page/model/cms-component-data';
-import { NavigationComponentService } from '../../../lib/cms-lib/navigation/navigation.component.service';
 import { GenericLinkModule } from '../../../lib/ui/components/generic-link/generic-link.module';
+import { NavigationComponentService } from '../navigation/navigation.component.service';
 import { FooterNavigationComponent } from './footer-navigation.component';
 
 @NgModule({
