+++ projects/storefrontlib/src/cms-components/navigation/index.ts
@@ -1,9 +1,15 @@
 // breadcrumb
 export * from './breadcrumb/breadcrumb.component';
 export * from './breadcrumb/breadcrumb.module';
+// category navigation
+export * from './category-navigation/category-navigation.component';
+export * from './category-navigation/category-navigation.module';
 // footer
 export * from './footer-navigation/footer-navigation.component';
 export * from './footer-navigation/footer-navigation.module';
+// navigation
+export * from './navigation/navigation.component';
+export * from './navigation/navigation.module';
 // searchbox
 export * from './search-box/search-box-component.service';
 export * from './search-box/search-box.component';
