+++ projects/storefrontlib/src/lib/cms-lib/cms-lib.module.ts
@@ -1,10 +1,15 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { CartComponentModule } from '../../cms-components/checkout/cart/cart.module';
-// import { BannerModule } from '../../cms-components/content/banner/banner.module';
+import { BannerModule } from '../../cms-components/content/banner/banner.module';
 import { SiteContextSelectorModule } from '../../cms-components/misc/site-context-selector/site-context-selector.module';
-import { BreadcrumbModule } from '../../cms-components/navigation/breadcrumb/breadcrumb.module';
-import { FooterNavigationModule } from '../../cms-components/navigation/footer-navigation/footer-navigation.module';
+import {
+  BreadcrumbModule,
+  CategoryNavigationModule,
+  FooterNavigationModule,
+  SearchBoxModule,
+} from '../../cms-components/navigation/index';
+import { NavigationModule } from '../../cms-components/navigation/navigation/navigation.module';
 import { HamburgerMenuModule, SkipLinkModule } from '../../layout/index';
 import { CloseAccountModule } from '../my-account/close-account/close-account.module';
 import { OrderDetailsModule } from '../my-account/order/order-details/order-details.module';
@@ -16,12 +21,9 @@ import { UpdateProfileModule } from '../my-account/update-profile/update-profile
 import { ProductListModule } from '../product/components/product-list/product-list.module';
 import { ProductTabsModule } from '../product/components/product-tabs/product-tabs.module';
 import { AddressBookModule } from './address-book/address-book.module';
-import { CategoryNavigationModule } from './category-navigation/category-navigation.module';
 import { LinkModule } from './link/link.module';
-import { NavigationModule } from './navigation/navigation.module';
 import { CmsParagraphModule } from './paragraph/paragraph.module';
 import { ProductCarouselModule } from './product-carousel/product-carousel.module';
-import { SearchBoxModule } from './search-box/search-box.module';
 
 @NgModule({
   imports: [
@@ -30,7 +32,7 @@ import { SearchBoxModule } from './search-box/search-box.module';
     HamburgerMenuModule,
     CmsParagraphModule,
     LinkModule,
-    // BannerModule,
+    BannerModule,
     CategoryNavigationModule,
     NavigationModule,
     FooterNavigationModule,
