+++ projects/storefrontlib/src/lib/cms-lib/index.ts
@@ -1,10 +1,6 @@
-export * from './category-navigation/category-navigation.component';
-export * from './category-navigation/category-navigation.module';
 export * from './cms-lib.module';
 export * from './link/link.component';
 export * from './link/link.module';
-export * from './navigation/navigation.component';
-export * from './navigation/navigation.module';
 export * from './paragraph/paragraph.component';
 export * from './paragraph/paragraph.module';
 export * from './product-carousel/product-carousel.component';
